requirements.txt
discord.py
feedparser
requests
aiohttp
xai-sdk